requirements.txt
flask
google-genai
gTTS
gunicorn
pydub